Refactored a SharedProfileHeaderViewController, and included in Home and Keys VC.
